==> 0_cog-sparql-recuperation-des-donnees.ipynb <==
from SPARQLWrapper import SPARQLWrapper, JSON

def get_all_ids_from_object(url):
    objs = []

    sparql = SPARQLWrapper("http://rdf.insee.fr/sparql")
    query = """
    SELECT * WHERE {
     ?s a <"""+url+""">
    }
    """
    sparql.setQuery(query)
    sparql.setReturnFormat(JSON)
    results = sparql.query().convert()

    for result in results["results"]["bindings"]:
        objs.append(result['s']['value'])
    print('Nombre de {} : {}'.format(url.replace('http://rdf.insee.fr/def/geo#',''),len(objs)))
    return objs


coms = get_all_ids_from_object("http://rdf.insee.fr/def/geo#Commune")
comas = get_all_ids_from_object("http://rdf.insee.fr/def/geo#CommuneAssociee")
comds = get_all_ids_from_object("http://rdf.insee.fr/def/geo#CommuneDeleguee")
comfs = get_all_ids_from_object("http://rdf.insee.fr/def/geo#CommuneFusionee")
evt = get_all_ids_from_object("http://rdf.insee.fr/def/geo#EvenementGeographique")


def get_info_com(typeCom, comArray):
    print("-----"+typeCom+"-----")
    i = 0
    for com in comArray:
        if(i%100 == 0):
            print(i)
        i = i + 1
        sparql = SPARQLWrapper("http://rdf.insee.fr/sparql")
        sparql.setQuery("""
           DESCRIBE <{}>
        """.format(com))
        sparql.setReturnFormat(JSON)
        results = sparql.query().convert()
        newcom = {}

        for result in results["results"]["bindings"]:
            if(result['subject']['value'] == com):
                if(result['predicate']['value'] == 'http://rdf.insee.fr/def/geo#codeINSEE'):
                    newcom['codeInsee'] = result['object']['value']
                if(result['predicate']['value'] == 'http://rdf.insee.fr/def/geo#nom'):
                    newcom['nom'] = result['object']['value']
                if(result['predicate']['value'] == 'http://rdf.insee.fr/def/geo#nomSansArticle'):
                    newcom['nomSansArticle'] = result['object']['value']
            if(result['object']['value'] == com):
                if(result['predicate']['value'] == 'http://rdf.insee.fr/def/geo#creation'):
                    newcom['evtCreation'] = result['subject']['value'].replace('http://id.insee.fr/geo/evenementGeographique/','')
                if(result['predicate']['value'] == 'http://rdf.insee.fr/def/geo#suppression'):
                    newcom['evtSuppression'] = result['subject']['value'].replace('http://id.insee.fr/geo/evenementGeographique/','')

        newcom['id'] = com.replace('http://id.insee.fr/geo/commune/','').replace('http://id.insee.fr/geo/communeAssociee/','').replace('http://id.insee.fr/geo/communeDeleguee/','').replace('http://id.insee.fr/geo/communeFusionnee/','')
        newcom['type'] = typeCom

        listallcoms.append(newcom)

Récupération du détail de toutes les communes (de tout type COMA COMD COM COMF) dans un tableau

listallcoms = []
get_info_com('COM',coms)
get_info_com('COMA',comas)
get_info_com('COMD',comds)
get_info_com('COMF',comfs)

Sauvegarde dans un fichier local (pour éviter de rejouer ce code à chaque fois)

import json
with open('coms.json', 'w') as f:
    json.dump(listallcoms, f, indent=4)

Récupération du détail de tous les événements dans un tableau

listallevt = []
i = 0
for e in evt:
    if(i%100 == 0):
        print(i)
    i = i + 1
    sparql = SPARQLWrapper("http://rdf.insee.fr/sparql")
    sparql.setQuery("""
       DESCRIBE <{}>
    """.format(e))
    sparql.setReturnFormat(JSON)
    results = sparql.query().convert()
    newevt = {}
    newevt['creation'] = []
    newevt['suppression'] = []
    
    for result in results["results"]["bindings"]:
        if(result['predicate']['value'] == 'http://rdf.insee.fr/def/geo#creation'):
            newevt['creation'].append(result['object']['value'].split('/')[-1])
        if(result['predicate']['value'] == 'http://rdf.insee.fr/def/geo#suppression'):
            newevt['suppression'].append(result['object']['value'].split('/')[-1])
        if(result['predicate']['value'] == 'http://rdf.insee.fr/def/geo#date'):
            newevt['date'] = result['object']['value']
                
    newevt['id'] = e.replace('http://id.insee.fr/geo/evenementGeographique/','')
    newevt['type'] = 'evenement'

    listallevt.append(newevt)

Sauvegarde dans un fichier json pour éviter de rejouer le traitement

import json
with open('evts.json', 'w') as f:
    json.dump(listallevt, f, indent=4)
