==> 1_enrichissement_fichiers_json.ipynb <==
import pandas as pd
import json

On récupère les fichiers générés à l'étape 0

with open('coms.json') as json_file:
    coms = json.load(json_file)
with open('evts.json') as json_file:
    evts = json.load(json_file)

dfcom = pd.DataFrame(coms)
dfevt = pd.DataFrame(evts)

dfcom

dfevt

Enrichissement du dataframe événement en ajoutant le détail de chaque commune ajoutée ou supprimée dans deux nouvelles colonnes

def getComs(coms):
    arr = []
    for com in coms:
        mydict = dfcom[dfcom['id'] == com].to_dict(orient = 'records')
        if(len(mydict) > 0):
            arr.append(dfcom[dfcom['id'] == com].to_dict(orient = 'records')[0])
    return arr

dfevt['creationObj'] = dfevt['creation'].apply(lambda x: getComs(x))

dfevt['suppressionObj'] = dfevt['suppression'].apply(lambda x: getComs(x))

dfevt

Sauvegarde dans un fichier (le traitement est long donc inutile de le rejouer à chaque fois)

import json
with open('evts-detail.json', 'w') as f:
    json.dump(dfevt.to_dict(orient="records"), f, indent=4)
