==> 2_analysis.ipynb <==
import pandas as pd
import json

with open('coms-clean.json') as json_file:
    coms = json.load(json_file)
with open('evts-detail.json') as json_file:
    evts = json.load(json_file)

dfcom = pd.DataFrame(coms)
dfevt = pd.DataFrame(evts)

### Nombre de communes actuelles

Total

dfcom[dfcom['evtSuppression'].isna()].shape

Par type

dfcom[dfcom['evtSuppression'].isna()]['type'].value_counts()

Aucune commune sans événement de création

dfcom[dfcom['evtCreation'].isna()]

Comparaison fichier csv COG datagouv et données sparql

df = pd.read_csv('commune2021.csv')

df[(df['TYPECOM'] == 'COMA') & (df['COM'] == '19267')]

Filtre hors ARM : arrondissement

listcog = df[df['TYPECOM'] != 'ARM']['COM'].unique().tolist()

Filtre commune qui ont eu une suppression pour ne garder que les actuelles

listrdf = dfcom[dfcom['evtSuppression'].isna()]['codeInsee'].unique().tolist()

Comparaison des deux listes

list(set(listcog) - set(listrdf))

len(listcog)

len(listrdf)

=> On est bon

### Fonctions

Fonctions récupération evt

def getEvtByCodeInsee(codeInsee):
    list_coms = dfcom[dfcom['codeInsee'] == codeInsee].id.to_list()
    print(list_coms)
    dftest = dfevt
    dftest['is_create'] = dftest['creation'].apply(lambda x: not set(list_coms).isdisjoint(x))
    df1 = dftest[dftest['is_create'] == True]
    dftest['is_supp'] = dftest['suppression'].apply(lambda x: not set(list_coms).isdisjoint(x))
    df2 = dftest[dftest['is_supp'] == True]
    df3 = pd.concat([df1,df2])
    return df3[~df3.index.duplicated(keep='first')].sort_values(by=['date'])

def getEvtByNom(nom):
    list_coms = dfcom[dfcom['nom'] == nom].id.to_list()
    print(list_coms)
    dftest = dfevt
    dftest['is_create'] = dftest['creation'].apply(lambda x: not set(list_coms).isdisjoint(x))
    df1 = dftest[dftest['is_create'] == True]
    dftest['is_supp'] = dftest['suppression'].apply(lambda x: not set(list_coms).isdisjoint(x))
    df2 = dftest[dftest['is_supp'] == True]
    df3 = pd.concat([df1,df2])
    return df3[~df3.index.duplicated(keep='first')].sort_values(by=['date'])

getEvtByCodeInsee('27032')

getEvtByNom('Collonges-lès-Premières')

Analyse nombre création / suppression par événement

dfevt['nb_creation'] = dfevt['creation'].apply(lambda x: len(x))
dfevt['nb_suppression'] = dfevt['suppression'].apply(lambda x: len(x))

Combien d'événement ont concerné plus d'une commune en création et plus d'une commune en suppression :

dfevt.shape

dfevt[(dfevt['nb_creation'] > 1) & (dfevt['nb_suppression'] > 1)].shape[0]

dfevt.nb_creation.value_counts()

dfevt.nb_suppression.value_counts()

Combien d'événement ont impliqués plus de communes créées que de communes supprimées (tout type de commune confondu) :

dfevt[(dfevt['nb_creation'] > 1) & (dfevt['nb_suppression'] > 1) & (dfevt['nb_creation'] >= dfevt['nb_suppression']+1)].shape[0]

On à toujours au max qu'une seule création en plus que de suppression :

dfevt[(dfevt['nb_creation'] > 1) & (dfevt['nb_suppression'] > 1) & (dfevt['nb_creation'] >= dfevt['nb_suppression']+2)].shape[0]

On va rechercher les événements qui ont créé plus de commune COM que de suppression de COM :
=> si il en existe, cela veut dire qu'on ne peut pas garantir l'intégrité territoriale dans le temps sur certaines communes

dftest = dfevt[(dfevt['nb_creation'] > 1) & (dfevt['nb_suppression'] > 1) & (dfevt['nb_creation'] >= dfevt['nb_suppression']+1)]

arr = []
for index, row in dftest.iterrows():
    nb_com_in = 0
    nb_coma_in = 0
    nb_comd_in = 0
    nb_com_out = 0
    nb_coma_out = 0
    nb_comd_out = 0
    for com in row['creationObj']:
        if(com['type'] == 'COM'):
            nb_com_out = nb_com_out + 1
        if(com['type'] == 'COMD'):
            nb_comd_out = nb_comd_out + 1
        if(com['type'] == 'COMA'):
            nb_coma_out = nb_coma_out + 1
    for com in row['suppressionObj']:
        if(com['type'] == 'COM'):
            nb_com_in = nb_com_in + 1
        if(com['type'] == 'COMD'):
            nb_comd_in = nb_comd_in + 1
        if(com['type'] == 'COMA'):
            nb_coma_in = nb_coma_in + 1
    if(nb_com_out > 1):
        arr.append(row)
        print('---------')
        print('Evenement '+row['id']+' : ')
        print('Suppression : '+str(nb_com_in)+' communes, '+str(nb_comd_in)+' communes déléguées, '+str(nb_coma_in)+' communes associées')
        print('Creation : '+str(nb_com_out)+' communes, '+str(nb_comd_out)+' communes déléguées, '+str(nb_coma_out)+' communes associées')

dfout = pd.DataFrame(arr)

for index, row in dfout.iterrows():
    print('-------')
    mystr = ''
    for com in row['suppressionObj']:
        mystr = mystr+com['nom']+' ('+com['codeInsee']+' - '+com['type']+') ; '
    
    mystr = mystr + '\ndeviennent\n'
                                          
    for com in row['creationObj']:
        mystr = mystr+com['nom']+' ('+com['codeInsee']+' - '+com['type']+') ; '
    print(mystr)

arr = []
for index, row in dftest.iterrows():
    nb_com_in = 0
    nb_coma_in = 0
    nb_comd_in = 0
    nb_com_out = 0
    nb_coma_out = 0
    nb_comd_out = 0
    for com in row['creationObj']:
        if(com['type'] == 'COM'):
            nb_com_out = nb_com_out + 1
        if(com['type'] == 'COMD'):
            nb_comd_out = nb_comd_out + 1
        if(com['type'] == 'COMA'):
            nb_coma_out = nb_coma_out + 1
    for com in row['suppressionObj']:
        if(com['type'] == 'COM'):
            nb_com_in = nb_com_in + 1
        if(com['type'] == 'COMD'):
            nb_comd_in = nb_comd_in + 1
        if(com['type'] == 'COMA'):
            nb_coma_in = nb_coma_in + 1
    if(nb_coma_out > 1):
        arr.append(row)
        print('---------')
        print('Evenement '+row['id']+' : ')
        print('Suppression : '+str(nb_com_in)+' communes, '+str(nb_comd_in)+' communes déléguées, '+str(nb_coma_in)+' communes associées')
        print('Creation : '+str(nb_com_out)+' communes, '+str(nb_comd_out)+' communes déléguées, '+str(nb_coma_out)+' communes associées')

dfout = pd.DataFrame(arr)

for index, row in dfout.iterrows():
    print('-------')
    mystr = ''
    for com in row['suppressionObj']:
        mystr = mystr+com['nom']+' ('+com['codeInsee']+' - '+com['type']+') ; '
    
    mystr = mystr + '\ndeviennent\n'
                                          
    for com in row['creationObj']:
        mystr = mystr+com['nom']+' ('+com['codeInsee']+' - '+com['type']+') ; '
    print(mystr)

Evenement qui ont créé plus de 2 communes COM

arr = []
for index, row in dfevt.iterrows():
    nb_com = 0
    for com in row['creationObj']:
        if(com['type'] == 'COM'):
            nb_com = nb_com + 1
    if(nb_com > 1):
        arr.append(row)
dfout = pd.DataFrame(arr)

dfout.shape[0]

for index, row in dfout.iterrows():
    comcreate = [x['nomSansArticle'] for x in row['creationObj']]
    comsupp = [x['nomSansArticle'] for x in row['suppressionObj']]

    #print(comcreate)
    #print(comsupp)
    #print('---')


df = dfevt.sort_values(by=['date'])
df = df[~((df['creationObj'].str.len() == 0) & (df['suppressionObj'].str.len() == 0))]

df.head()

arr = []
cpt = 0
for index, row in df.iterrows():
    # 1 ou N création, 0 suppression
    if((len(row['creationObj']) >= 1) & (len(row['suppressionObj']) == 0)):
        for com in row['creationObj']:
            mydict = {}
            mydict['nom'] = com['nom']
            mydict['type'] = com['type']
            mydict['dateDebut'] = row['date']
            mydict['code'] = com['codeInsee']
            mydict['predecesseur'] = None
            mydict['id'] = com['type'] + '-' + com['codeInsee'] + '@' + row['date']
            mydict['idEvt'] = row['id']
            arr.append(mydict)
            
    # 1 ou N création, 1 suppression
    if((len(row['creationObj']) >= 1) & (len(row['suppressionObj']) == 1)):
        for com in row['creationObj']:
            mydict = {}
            mydict['nom'] = com['nom']
            mydict['type'] = com['type']
            mydict['dateDebut'] = row['date']
            mydict['code'] = com['codeInsee']
            mydict['predecesseur'] = row['suppressionObj'][0]['type'] + '-' + row['suppressionObj'][0]['codeInsee'] + '@' + df[df['id'] == row['suppressionObj'][0]['evtCreation']].iloc[0]['date']
            mydict['id'] = com['type'] + '-' + com['codeInsee'] + '@' + row['date']
            mydict['idEvt'] = row['id']
            arr.append(mydict)
            
    # N création, N suppression
    if((len(row['creationObj']) >= 1) & (len(row['suppressionObj']) > 1)):
        nbcom = 0
        codecom = ''
        datecom = ''
        for com in row['suppressionObj']:
            if(com['type'] == 'COM'):
                nbcom += 1
                codecom = com['codeInsee']
                datecom = df[df['id'] == com['evtCreation']].iloc[0]['date']
        if(nbcom > 1):
            for com in row['creationObj']:
                mydict = {}
                mydict['nom'] = com['nom']
                mydict['type'] = com['type']
                mydict['dateDebut'] = row['date']
                mydict['code'] = com['codeInsee']
                codesInsee = []
                for com2 in row['suppressionObj']:
                    codesInsee.append(com2['codeInsee'])
                codesInsee = list(dict.fromkeys(codesInsee))
                mydict['predecesseur'] = 'COMT-' + '-'.join(codesInsee) + '@' + row['date']
                mydict['id'] = com['type'] + '-' + com['codeInsee'] + '@' + row['date']
                mydict['idEvt'] = row['id']
                arr.append(mydict)
        elif(nbcom == 1):
            for com in row['creationObj']:
                mydict = {}
                mydict['nom'] = com['nom']
                mydict['type'] = com['type']
                mydict['dateDebut'] = row['date']
                mydict['code'] = com['codeInsee']
                mydict['predecesseur'] = 'COM-' + codecom + '@' + datecom
                mydict['id'] = com['type'] + '-' + com['codeInsee'] + '@' + row['date']
                mydict['idEvt'] = row['id']
                arr.append(mydict)
        else:
            print('pb en perspective - 0 commune supprimée mais des comd et coma')

    # 0 création, N suppression
print(cpt)
    
    

len(arr)

toto = pd.DataFrame(arr)
toto['successeur'] = ''

arr = []
cpt = 0
for index, row in df.iterrows():
    
    # 1 ou N suppression, 0 création
    if((len(row['suppressionObj']) >= 1) & (len(row['creationObj']) == 0)):
        for com in row['suppressionObj']:
            mydict = {}
            mydict['nom'] = com['nom']
            mydict['type'] = 'COMP'
            mydict['dateDebut'] = row['date']
            mydict['code'] = com['codeInsee']
            mydict['predecesseur'] = com['type'] + '-' + com['codeInsee'] + '@' + toto[toto['idEvt'] == com['evtCreation']].iloc[0]['dateDebut']
            mydict['id'] = 'COMP-' + com['codeInsee'] + '@' + row['date']
            mydict['idEvt'] = row['id']
            arr.append(mydict) 
            
            mydate = toto[(toto['dateDebut'] <= row['date']) & (toto['id'].str.contains(com['type']+'-'+com['codeInsee']))].sort_values(by='dateDebut',ascending=False).iloc[0]['dateDebut']
            toto.loc[(toto['id'] == com['type']+'-'+com['codeInsee']+'@'+mydate),'successeur'] = 'COMP-' + com['codeInsee'] + '@' + row['date']
            
            
    # N suppression, N creation
    if((len(row['suppressionObj']) > 1) & (len(row['creationObj']) >= 1)):
        cpt = cpt + 1
        if(cpt % 200 == 0):
            print(cpt)
        codesInsee = []
        for com in row['suppressionObj']:
            if(com['type'] == 'COM'):
                codesInsee.append(com['codeInsee'])
        codesInsee = list(dict.fromkeys(codesInsee))
        mydict = {}
        mydict['nom'] = 'Commune Transitoire'
        mydict['type'] = 'COMT'
        mydict['dateDebut'] = row['date']
        mydict['code'] = '-'.join(codesInsee)
        mydict['id'] = 'COMT-' + '-'.join(codesInsee) + '@' + row['date']
        mydict['idEvt'] = row['id']
        mydict['predecesseur'] = None
        arr.append(mydict)
        for com in row['suppressionObj']:
            try:
                mydate = toto[(toto['dateDebut'] <= row['date']) & (toto['id'].str.contains(com['type']+'-'+com['codeInsee']))].sort_values(by='dateDebut',ascending=False).iloc[0]['dateDebut']
                toto.loc[(toto['id'] == com['type']+'-'+com['codeInsee']+'@'+mydate),'successeur'] = 'COMT-' + '-'.join(codesInsee) + '@' + row['date']
            except:
                print('Pb intégrité base SPARQL - id Evt INSEE : '+row['id'])
    
    if((len(row['suppressionObj']) == 1) & (len(row['creationObj']) == 1)):
        mydate = toto[(toto['dateDebut'] <= row['date']) & (toto['id'].str.contains(row['suppressionObj'][0]['type']+'-'+row['suppressionObj'][0]['codeInsee']))].sort_values(by='dateDebut',ascending=False).iloc[0]['dateDebut']
        toto.loc[(toto['id'] == row['suppressionObj'][0]['type']+'-'+row['suppressionObj'][0]['codeInsee']+'@'+mydate),'successeur'] = row['creationObj'][0]['type'] + '-' + row['creationObj'][0]['codeInsee'] + '@' + row['date']
        
    
print(cpt)


tutu = pd.DataFrame(arr)

titi = pd.concat([toto,tutu])

titi['successeur'] = titi['successeur'].apply(lambda x: None if x == '' else x)

titi[titi['successeur'].isna()].shape

titi[titi['nom'] == 'Vayres-sur-Essonne']

for index, row in titi.iterrows():
    if(row['successeur'] is None):
        if(titi[titi['predecesseur'] == row['id']].shape[0] > 0):
            row['successeur'] = titi[titi['predecesseur'] == row['id']].iloc[0]['id']

for index, row in titi.iterrows():
    if(row['id'] == row['successeur']):
        row['successeur'] = None

cog = pd.read_csv('commune2021.csv')
cog = cog[cog['TYPECOM'] != 'ARM']

out = titi.copy()

out = out[out['successeur'].isna()]

cpt = 0
for index, row in out.iterrows():
    cpt = cpt + 1
    if(cpt % 1000 == 0):
        print(cpt)
    if((row['type'] != 'COMT') & (row['type'] != 'COMP')):
        if(cog[(cog['TYPECOM'] == row['type']) & (cog['COM'] == row['code'])].shape[0] == 0):
            print(row)

cpt = 0
for index, row in cog.iterrows():
    cpt = cpt + 1
    if(cpt % 1000 == 0):
        print(cpt)
    if(out[(out['type'] == row['TYPECOM']) & (out['code'] == row['COM'])].shape[0] == 0):
        print(row)

import json
with open('graph-coms.json', 'w') as f:
    json.dump(titi.to_dict(orient="records"), f, indent=4)

cog

cog[cog['NCCENR'] == 'Boissy-sur-Damville']

titi[titi['code'] == '27080']

titi[titi['code'].str.contains('27416')]

dfevt[dfevt['id'] == '904dcf14-b895-4686-a5c5-7304c6f63cae']

Chaque commune possède un identifiant unique : TYPE-CODEINSEE@DATE_EVT_CREATION

Exemple :
Tartampion 27455 Tartampion 27455 le 1er janvier 1970
a considéré comme deux objets distincts même si le cas semble bizarre, du point de vue de la base de données, il faut le considérer comme deux objets distincts. Ici, avec deux identifiants :
COM-27455@1943-01-01 
COM-27455@1970-01-01

Pour gérer l'historique et le graphe, nous avons besoin de créer deux nouveaux types de commune :
- COMP : commune Périmée : une fois qu'une commune est supprimée mais n'a pas de successeur, il faut créer une COMP
- COMT : commune Transitoire : lorsque l'événement ne permet pas d'associer un successeur ou un prédecesseur unique, nous avons besoin de créer une commune transitoire qui se charge de faire la jointure entre les communes supprimées et communes créées

Quand on parle ci-dessous de création ou de suppression, il s'agit de création ou de suppression d'objet Commune COM (et non pas COMA ou COMD) :

Gestion des événements, plusieurs cas :
- 0 suppression, 1 création : 
    - la commune créé ne possède pas de prédécesseur
- 0 suppression, N création : ok
    - les communes créées ne possèdent pas de prédecesseurs
- 1 suppression, 1 création : ok
    - la commune créé possède un prédécesseur
    - la commune supprimée possède un successeur
- 1 suppression, N création : besoin d'une commune transitoire
    - la commune supprimée a pour successeur la commune transitoire
    - les communes créées ont pour prédecesseur la commune transitoire
    - la commune transitoire créée n'a pas de successeur ni de prédecesseur
- N suppression, 0 création : ok
    - les communes supprimées ont pour chacune un objet COMP (Commune Périmée) a créer.
- N suppression, 1 création : besoin d'une commune transitoire
    - les communes supprimées ont pour successeurs la commune transitoire
    - la commune créée a pour prédecesseur la commune transitoire
    - la commune transitoire créée n'a pas de successeur ni de prédecesseur
- N suppression, N création : besoin d'une commune transitoire
    - les communes supprimées ont pour successeurs la commune transitoire
    - les communes créées ont pour prédecesseurs la commune transitoire
    - la commune transitoire créée n'a pas de successeur ni de prédecesseur

Propriétés des objets communes COM, COMD... :
- nom ; type ; dateDebut ; code ; predecesseur ; id ; idEvt ; successeur
Propriétés des objets commune transitoire COMT :
- nom ; type ; dateDebut ; code ; id ; idEvt


Cas particulier:
- Hauteville https://rdf.insee.fr/sparql/describe?uri=http%3A%2F%2Fid.insee.fr%2Fgeo%2Fcommune%2Fdef17c47-5128-401d-9e42-0690d8c0daae&default-graph-uri=data
- Créé en 1943, supprimé en 1942....
